+++ app/src/main/AndroidManifest.xml
@@ -182,7 +182,7 @@
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
 
-        <activity android:name=".src.main.seller.store.product.SellerProductAddActivity"
+        <activity android:name=".src.main.seller.store.productfeed.product.SellerProductAddActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
@@ -194,7 +194,7 @@
             android:theme="@style/AppTheme2">
         </activity>
 
-        <activity android:name=".src.main.seller.store.product.SellerProductEditActivity"
+        <activity android:name=".src.main.seller.store.productfeed.product.SellerProductEditActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme2">
