+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/SellerProductFeedFragment.kt
@@ -13,7 +13,7 @@ import com.codepatissier.keki.src.main.consumer.store.productfeed.ConsumerStoreP
 import com.codepatissier.keki.src.main.consumer.store.productfeed.ConsumerStoreProductFeedView
 import com.codepatissier.keki.src.main.consumer.store.productfeed.model.ConsumerStoreProductFeedResponse
 import com.codepatissier.keki.src.main.consumer.store.productfeed.productdetail.ConsumerStoreProductDetailFeedActivity
-import com.codepatissier.keki.src.main.seller.store.product.SellerProductAddActivity
+import com.codepatissier.keki.src.main.seller.store.productfeed.product.SellerProductAddActivity
 import com.codepatissier.keki.src.main.seller.store.productfeed.SellerProductDetailFeedActivity
 import com.codepatissier.keki.util.viewpager.storemain.StoreMainProductData
 import com.codepatissier.keki.util.viewpager.storemain.seller.SellerStoreMainProductAdapter
