+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/SellerProductFeedDetailRetrofitInterface.kt
@@ -2,8 +2,8 @@ package com.codepatissier.keki.src.main.seller.store.productfeed
 
 import com.codepatissier.keki.config.BaseResponse
 import com.codepatissier.keki.src.main.consumer.store.productfeed.productdetail.model.ConsumerStoreProductFeedDetailResponse
-import com.codepatissier.keki.src.main.seller.store.product.model.SellerProductAddBody
-import com.codepatissier.keki.src.main.seller.store.product.model.SellerProductAddResponse
+import com.codepatissier.keki.src.main.seller.store.productfeed.product.model.SellerProductAddBody
+import com.codepatissier.keki.src.main.seller.store.productfeed.product.model.SellerProductAddResponse
 import com.codepatissier.keki.src.main.seller.store.productfeed.model.SellerProductFeedDetailResponse
 import retrofit2.Call
 import retrofit2.http.*
