+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/SellerProductFeedDetailService.kt
@@ -2,7 +2,7 @@ package com.codepatissier.keki.src.main.seller.store.productfeed
 
 import com.codepatissier.keki.config.ApplicationClass
 import com.codepatissier.keki.config.BaseResponse
-import com.codepatissier.keki.src.main.seller.store.product.model.SellerProductAddBody
+import com.codepatissier.keki.src.main.seller.store.productfeed.product.model.SellerProductAddBody
 import com.codepatissier.keki.src.main.seller.store.productfeed.model.SellerProductFeedDetailResponse
 import retrofit2.Call
 import retrofit2.Callback
