+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/product/SellerProductAddActivity.kt
@@ -1,4 +1,4 @@
-package com.codepatissier.keki.src.main.seller.store.product
+package com.codepatissier.keki.src.main.seller.store.productfeed.product
 
 import android.content.Intent
 import android.net.Uri
@@ -12,13 +12,14 @@ import com.bumptech.glide.Glide
 import com.codepatissier.keki.R
 import com.codepatissier.keki.config.BaseActivity
 import com.codepatissier.keki.databinding.ActivitySellerProductAddAndEditBinding
-import com.codepatissier.keki.src.main.seller.store.product.model.SellerProductAddBody
-import com.codepatissier.keki.src.main.seller.store.product.model.SellerProductAddResponse
+import com.codepatissier.keki.src.main.seller.store.productfeed.product.model.SellerProductAddBody
+import com.codepatissier.keki.src.main.seller.store.productfeed.product.model.SellerProductAddResponse
 import com.google.firebase.storage.FirebaseStorage
 import java.text.SimpleDateFormat
 import java.util.*
 
-class SellerProductAddActivity:BaseActivity<ActivitySellerProductAddAndEditBinding>(ActivitySellerProductAddAndEditBinding::inflate), SellerProductAddView {
+class SellerProductAddActivity:BaseActivity<ActivitySellerProductAddAndEditBinding>(ActivitySellerProductAddAndEditBinding::inflate),
+    SellerProductAddView {
     private lateinit var launcher : ActivityResultLauncher<Intent>
     var ProfileUri : Uri?= null
     var fbStorage : FirebaseStorage?= null
