+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/product/model/SellerProductAddBody.kt
@@ -1,4 +1,4 @@
-package com.codepatissier.keki.src.main.seller.store.product.model
+package com.codepatissier.keki.src.main.seller.store.productfeed.product.model
 
 data class SellerProductAddBody(
     val dessertName: String,
